refactor: Use userID for bmi and tee requests
Requests for a user's bmi and tee have been refactored to use the userID as a request parameter.
That way the correct attributes needed to request BMI and TEE can be pulled from the user's info.

/user/bmi and /user/tee routes were updated to include the user id at the end: /user/bmi/{id}
Tests for bmi and tee have also been updated.

TODO next: Authenticate both endpoints