feat: Added authentication to /user/bmi and /user/tee endpoints
Endpoints to retrieve a user's bmi and tee were recently updated
to use the user's id as a request parameter to then pull the
user's info from the db based on that id.
Now, authentication has been added to both endpoints so that
the bmi and tee data cannot be retrieved for a user other than
the owner of the account.