feat: Imported client library for Activity Collector Service
The Activity Collector is a service responsible for receiving and storing
user's activity data. This commit includes work needed to use the Activity Collector's
client library to be able to retrieve a user's activity data.

So far there is only one endpoint in the client to retrieve a data point. There will
be more functionality added to the Activity Collector later, resulting in more uses for
the client library.